llvm: fix promoted allocas + lifetime.start intrinsics
Allocations should happen with the lifetime.start. Otherwise there may
be use-after-free bugs. E.g. when an alloca's lifetime is only valid
within a loop and the memory is freed at the end of the loop (due to a
lifetime.end intrinsic). The next iteration of the loop will try and use
the old memory, but this has obviously been freed :(